refactor: wallet clean up
